split ultimate.py into ultimate module with class submodules
