+++ charts/Mens_Advancement_Probabilities_v1.png
