+++ charts/Womens_Advancement_Probabilities_v1.png
