+++ charts/Womens_Advancement_Probabilities_v2.png
