+++ requirements.txt
@@ -1,0 +1,6 @@
+pandas==1.0.1
+numpy==1.18.2
+matplotlib==3.2.1
+scipy==1.4.1
+scikit-learn==0.22.2.post1
+seaborn==0.10.0
